Adding Singletone connection to database. It is working dude!

+++ src/main/java/com/jimenezict/bicing/getstationstatus/service/ParserToDatabaseImpl.java
@@ -32,6 +32,6 @@ public class ParserToDatabaseImpl implements ParserToDatabase {
     }
 
     private double usageCalculation(Station station) {
-        return (station.getNum_bikes_available()) / (station.getNum_bikes_available() + station.getNum_docks_available());
+        return (double)(station.getNum_bikes_available()) / (station.getNum_bikes_available() + station.getNum_docks_available());
     }
 }
